Update AbstractMovementTests and CheckTests
