+++ src/main/java/com/chess/backend/gamemodel/abstractmoves/MoveForward.java
@@ -1,7 +1,7 @@
 package com.chess.backend.gamemodel.abstractmoves;
 
 import com.chess.backend.gamemodel.*;
-import com.chess.backend.gamemodel.pieces.Piece;
+import com.chess.backend.gamemodel.Piece;
 import com.chess.backend.services.ChessboardService;
 
 import java.util.HashSet;
