+++ src/main/java/com/chess/backend/gamemodel/abstractmoves/MoveOneRight.java
@@ -2,8 +2,7 @@ package com.chess.backend.gamemodel.abstractmoves;
 
 import com.chess.backend.gamemodel.Chessboard;
 import com.chess.backend.gamemodel.Move;
-import com.chess.backend.gamemodel.Square;
-import com.chess.backend.gamemodel.pieces.Piece;
+import com.chess.backend.gamemodel.Piece;
 
 import java.util.Set;
