+++ src/main/java/com/chess/backend/gamemodel/abstractmoves/MovePawnCaptureBackward.java
@@ -1,7 +1,7 @@
 package com.chess.backend.gamemodel.abstractmoves;
 
 import com.chess.backend.gamemodel.*;
-import com.chess.backend.gamemodel.pieces.Piece;
+import com.chess.backend.gamemodel.Piece;
 
 import java.util.HashSet;
 import java.util.Set;
