--- src/main/java/com/chess/backend/gamemodel/pieces/Ferz.java
@@ -1,102 +1,0 @@
-package com.chess.backend.gamemodel.pieces;
-
-import com.chess.backend.gamemodel.*;
-import com.chess.backend.gamemodel.abstractmoves.*;
-import com.chess.backend.gamemodel.constants.Color;
-import com.chess.backend.gamemodel.constants.PieceType;
-import lombok.Data;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Set;
-
-@Data
-public class Ferz extends Piece {
-    //    private Square square;
-    private  Integer posX;
-    private  Integer posY;
-    private  Player player;
-    private  PieceType type = PieceType.FERZ;
-    private  boolean motioned;
-    private  final boolean clockwise;
-
-    public Ferz(Player player, boolean clockwise) {
-        this.player = player;
-        this.clockwise = clockwise;
-    }
-
-    @Override
-    public HashSet<Move> getAllowedFullMoves(Chessboard chessboard) {
-        HashSet<Move> allowedMoves = new HashSet<>();
-        allowedMoves.addAll(MoveOneDiagonal.concretise(chessboard, this, true, false, true));
-        return allowedMoves;
-    }
-
-    @Override
-    public ArrayList<Square> getAllowedMoves(Chessboard chessboard) {
-        Set<Move> allowedFullMoves = getAllowedFullMoves(chessboard);
-        ArrayList<Square> allowedMoves = new ArrayList<>();
-
-        for (Move move :
-                allowedFullMoves) {
-            allowedMoves.add(move.getTo());
-        }
-
-        return allowedMoves;
-    }
-
-    @Override
-    public boolean getMainDirection() {
-        return this.clockwise;
-    }
-
-    @Override
-    public PieceType getType() {
-        return this.type;
-    }
-
-    @Override
-    public void setType(PieceType type) {
-        this.type = type;
-    }
-
-    @Override
-    public Color getColor() {
-        return this.player.getColor();
-    }
-
-    @Override
-    public String getSymbol() {
-        return this.getType().getSymbol();
-    }
-
-    @Override
-    public boolean isMotioned() {
-        return motioned;
-    }
-
-    @Override
-    public void setMotioned(boolean motioned) {
-        this.motioned = motioned;
-    }
-
-    @Override
-    public Player getPlayer() {
-        return player;
-    }
-
-    @Override
-    public boolean isClockwise() {
-        return clockwise;
-    }
-
-    @Override
-    public String toString() {
-        return "Piece{" +
-                "type=" + type +
-                ", player=" + player +
-                ", motioned=" + motioned +
-                ", clockwise=" + clockwise +
-                '}';
-    }
-}
