+++ src/main/java/com/chess/backend/repository/GameRepository.java
@@ -1,20 +1,12 @@
 package com.chess.backend.repository;
 
-import com.chess.backend.domain.models.IPiece;
 import com.chess.backend.domain.repository.IGameRepository;
 import com.chess.backend.gamemodel.ChessGame;
-import com.chess.backend.gamemodel.Square;
+import com.chess.backend.gamemodel.Piece;
 import com.chess.backend.gamemodel.constants.PieceType;
-import com.chess.backend.gamemodel.pieces.*;
-import com.chess.backend.restController.objects.ChessboardObject;
-import com.chess.backend.restController.objects.SquareObject;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.GsonBuilder;
-import com.google.gson.TypeAdapter;
-import com.google.gson.TypeAdapterFactory;
-import com.google.gson.reflect.TypeToken;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import com.google.gson.Gson;
@@ -24,7 +16,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
 
 
 @Service
@@ -59,35 +50,7 @@ public class GameRepository implements IGameRepository {
 
                     if (piece != null) {
                         String pieceJson = new Gson().toJson(piece);
-                        Class pieceClass = null;
-                        if (piece.get("type").toString().equals(PieceType.PAWN.getLabel().toUpperCase())) {
-                            pieceClass = Pawn.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.ROOK.getLabel().toUpperCase())) {
-                            pieceClass = Rook.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.KNIGHT.getLabel().toUpperCase())) {
-                            pieceClass = Knight.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.BISHOP.getLabel().toUpperCase())) {
-                            pieceClass = Bishop.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.QUEEN.getLabel().toUpperCase())) {
-                            pieceClass = Queen.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.KING.getLabel().toUpperCase())) {
-                            pieceClass = King.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.CANNON.getLabel().toUpperCase())) {
-                            pieceClass = Cannon.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.FERZ.getLabel().toUpperCase())) {
-                            pieceClass = Ferz.class;
-                        }
-                        else if (piece.get("type").toString().equals(PieceType.WAZIR.getLabel().toUpperCase())) {
-                            pieceClass = Wazir.class;
-                        }
-                        Piece pawn=  new Gson().fromJson(pieceJson, (Type) pieceClass);
+                        Piece pawn = new Gson().fromJson(pieceJson, Piece.class);
                         System.out.println();
                         newGame.getChessboard().getSquares().get(i).get(j).setPiece(pawn);
                     }
