+++ src/main/java/com/chess/backend/services/ChessGameService.java
@@ -2,18 +2,12 @@ package com.chess.backend.services;
 
 import com.chess.backend.domain.models.IGame;
 import com.chess.backend.domain.repository.IGameRepository;
-import com.chess.backend.domain.services.IGameService;
-import com.chess.backend.domain.services.INewGameService;
 import com.chess.backend.gamemodel.*;
-import com.chess.backend.domain.models.IPiece;
 import com.chess.backend.gamemodel.constants.Event;
 import com.chess.backend.gamemodel.constants.PieceType;
-import com.chess.backend.gamemodel.pieces.Piece;
-import com.chess.backend.repository.GameRepository;
-import lombok.Data;
+import com.chess.backend.gamemodel.Piece;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
